test: stabilize remote state journeys
